// src/types/beats/cosmos.ts
// Auto-generated by scripts/generate-block-types.ts
// Generated: 2026-03-17
// Sources:
//   https://jiti.indexing.co/networks/cosmos/30234630
//   https://jiti.indexing.co/networks/cosmos/30234530
//   https://jiti.indexing.co/networks/cosmos/30233630
// Samples: 3
export interface CosmosBlockBlockIdParts {
  total: number;
  hash: string;
}

export interface CosmosBlockBlockId {
  hash: string;
  parts: CosmosBlockBlockIdParts;
}

export interface CosmosBlockBlockHeaderVersion {
  block: string;
}

export interface CosmosBlockBlockHeaderLastBlockIdParts {
  total: number;
  hash: string;
}

export interface CosmosBlockBlockHeaderLastBlockId {
  hash: string;
  parts: CosmosBlockBlockHeaderLastBlockIdParts;
}

export interface CosmosBlockBlockHeader {
  version: CosmosBlockBlockHeaderVersion;
  chain_id: string;
  height: string;
  time: string;
  last_block_id: CosmosBlockBlockHeaderLastBlockId;
  last_commit_hash: string;
  data_hash: string;
  validators_hash: string;
  next_validators_hash: string;
  consensus_hash: string;
  app_hash: string;
  last_results_hash: string;
  evidence_hash: string;
  proposer_address: string;
}

export interface CosmosBlockBlockData {
  txs: string[];
}

export interface CosmosBlockBlockEvidence {
  evidence: unknown[];
}

export interface CosmosBlockBlockLastCommitBlockIdParts {
  total: number;
  hash: string;
}

export interface CosmosBlockBlockLastCommitBlockId {
  hash: string;
  parts: CosmosBlockBlockLastCommitBlockIdParts;
}

export interface CosmosBlockBlockLastCommitSignatures {
  block_id_flag: number;
  validator_address: string;
  timestamp: string;
  signature: string;
}

export interface CosmosBlockBlockLastCommit {
  height: string;
  round: number;
  block_id: CosmosBlockBlockLastCommitBlockId;
  signatures: CosmosBlockBlockLastCommitSignatures[];
}

export interface CosmosBlockBlock {
  header: CosmosBlockBlockHeader;
  data: CosmosBlockBlockData;
  evidence: CosmosBlockBlockEvidence;
  last_commit: CosmosBlockBlockLastCommit;
}

export interface CosmosBlockTxsResultsEventsAttributes {
  key: string;
  value: string;
  index: boolean;
}

export interface CosmosBlockTxsResultsEvents {
  type: string;
  attributes: CosmosBlockTxsResultsEventsAttributes[];
}

export interface CosmosBlockTxsResults {
  code: number;
  data: string;
  log: string;
  info: string;
  gas_wanted: string;
  gas_used: string;
  events: CosmosBlockTxsResultsEvents[];
  codespace: string;
}

export interface CosmosBlockFinalizeBlockEventsAttributes {
  key: string;
  value: string;
  index: boolean;
}

export interface CosmosBlockFinalizeBlockEvents {
  type: string;
  attributes: CosmosBlockFinalizeBlockEventsAttributes[];
}

export interface CosmosBlockConsensusParamUpdatesBlock {
  max_bytes: string;
  max_gas: string;
}

export interface CosmosBlockConsensusParamUpdatesEvidence {
  max_age_num_blocks: string;
  max_age_duration: string;
  max_bytes: string;
}

export interface CosmosBlockConsensusParamUpdatesValidator {
  pub_key_types: string[];
}

export interface CosmosBlockConsensusParamUpdatesVersion {
}

export interface CosmosBlockConsensusParamUpdatesAbci {
}

export interface CosmosBlockConsensusParamUpdates {
  block: CosmosBlockConsensusParamUpdatesBlock;
  evidence: CosmosBlockConsensusParamUpdatesEvidence;
  validator: CosmosBlockConsensusParamUpdatesValidator;
  version: CosmosBlockConsensusParamUpdatesVersion;
  abci: CosmosBlockConsensusParamUpdatesAbci;
}

export interface CosmosBlock {
  block_id: CosmosBlockBlockId;
  block: CosmosBlockBlock;
  height: string;
  txs_results: CosmosBlockTxsResults[];
  finalize_block_events: CosmosBlockFinalizeBlockEvents[];
  validator_updates: null;
  consensus_param_updates: CosmosBlockConsensusParamUpdates;
  app_hash: string;
  _network: string;
  _vm: string;
}